grafo-arvore-projeto/backend/requirements.txt
Django>=4.0,<5.0
django-cors-headers>=3.10.0
djangorestframework>=3.12.0
